refactoring : only one message Msg::DataFetched used by fetch_json, instead of 4
There is now only one message ***Msg::DataFetched*** used by fetch_json (instead of 4).
However we still need to distinguish fetched data by type, regarding the request (the structure to be passed on fetch_json has to be specific to the data received for the JSON to be deserialized in this structure).
For that purpose, the ***Msg::DataFetched*** accepts an ***enum FetchedData*** (4 available values : Balance, MarketChannels, Volume or Impressions).