Improve the fetch error management
we have to deal with the error directly when decoding the FethcObject (inspired from https://github.com/MartinKavik/seed-rs-realworld).
In that way, I’ve created a vector of ’errors’ in Model to push error(s), if any, when decoding.
Then, in ‘view’, the errors are parsed and some HTML elements are created in the page to display the errors.
I’ve used `<details> / <summary>` tags to display errors in a collapsible way.
I’ve also changed `view` a bit : 
 - before : if market_channels fetch fails, nothing is displayed
 - after : every information that has been really fetched is displayed, whereas some others are not, because of the error(s). In case an information can't be computed, 'N/A' is displayed instead of the value.